Django/03_bookmark/pyproject.toml
[project]
name = "oz-bookmark"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipython (>=9.0.2,<10.0.0)",
    "django-extensions (>=3.2.3,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
